added the thread services
